Refactor : List page - useInitWithLocalData, useSaveChanged
